requirements.txt
numpy>=1.14.3
torchvision==0.2.1
scipy>=1.1.0
gitpython==2.1.11
torchnet==0.0.4
pydot==1.2.4
tabulate==0.8.2
pandas>=0.22.0
jupyter>=1.0.0
matplotlib>=2.2.2
qgrid==1.0.2
graphviz==0.8.2
ipywidgets==7.1.2
bqplot==0.10.5
pyyaml
pytest==3.5.1
xlsxwriter>=1.1.1
pretrainedmodels
